+++ src/plugins/validate-z-schema.ts
@@ -34,7 +34,7 @@ function _getValidator(
     if (!VALIDATOR_CACHE.has(hash)) {
         const validator = new (ZSchema as any)();
         const validatorFun = (obj: any) => {
-            validator.validate(obj, rxSchema.jsonID);
+            validator.validate(obj, rxSchema.jsonSchema);
             return validator;
         };
         VALIDATOR_CACHE.set(hash, validatorFun);
@@ -68,7 +68,7 @@ const validate = function (
         throw newRxError('VD2', {
             errors: formattedZSchemaErrors,
             obj,
-            schema: this.jsonID
+            schema: this.jsonSchema
         });
     }
 };